LPS-189207 Rename connection provider class and move it to QuartzSchedulerEngine
