--- modules/apps/portal-scheduler/portal-scheduler-quartz/src/main/java/com/liferay/portal/scheduler/quartz/internal/QuartzConnectionProvider.java
@@ -1,56 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.scheduler.quartz.internal;
-
-import com.liferay.petra.lang.SafeCloseable;
-import com.liferay.portal.kernel.log.Log;
-import com.liferay.portal.kernel.log.LogFactoryUtil;
-import com.liferay.portal.kernel.security.auth.CompanyThreadLocal;
-import com.liferay.portal.kernel.util.InfrastructureUtil;
-
-import java.sql.Connection;
-
-import javax.sql.DataSource;
-
-import org.quartz.utils.ConnectionProvider;
-
-/**
- * @author Bruno Farache
- */
-public class QuartzConnectionProvider implements ConnectionProvider {
-
-	@Override
-	public Connection getConnection() {
-		Connection connection = null;
-
-		try (SafeCloseable safeCloseable =
-				CompanyThreadLocal.setWithSafeCloseable(_companyId)) {
-
-			DataSource dataSource = InfrastructureUtil.getDataSource();
-
-			connection = dataSource.getConnection();
-		}
-		catch (Exception exception) {
-			_log.error(exception);
-		}
-
-		return connection;
-	}
-
-	@Override
-	public void shutdown() {
-	}
-
-	protected QuartzConnectionProvider(long companyId) {
-		_companyId = companyId;
-	}
-
-	private static final Log _log = LogFactoryUtil.getLog(
-		QuartzConnectionProvider.class);
-
-	private final long _companyId;
-
-}
